requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.7.1+cpu  # Newest stable CPU version available
torchvision==0.17.1+cpu  # Matching version
flask==2.3.2
flask-cors==4.0.0
pillow==10.3.0
git+https://github.com/PramaLLC/BEN2.git
gunicorn==21.2.0